Added before and after comparison of shards for node
We have been seeing nodes return to the cluster but all of it's
shards it held go into init and zero remain on the restarted node.

Proceeding when green in this case could eventually run a cluster
out of disk space if the rate at which we push shards off exceeds
the rate at which shards return.  We prevent this scenario by
saving the original number of shards on the node and requiring that
66% of those shards are present before continuing.